lib: vsprintf: optimised put_dec() function
The put_dec() and family of functions were based on a code
optimised for processors with 8-bit ALU but since we don't
need to limit ourselves to such small ALUs, the code was
optimised and used capacities of an 16-bit ALU anyway.

This patch goes further and uses the full capacity of the
processor's ALU.

On 64-bit machines, the number is repeatedly divided by
100,000 to split it into 5-decimal-digit which are converted
using the obvious base conversion algorithm expect division by
ten is replaced with multiplication and shifts.

On 32-bit machines, no division is performed at all and in
particular, no 64-bit division is performed.  This can speed
up conversion a few times and up to 10 times!

Signed-off-by: Michal Nazarewicz <mina86@mina86.com>
Signed-off-by: Douglas W. Jones <jones@cs.uiowa.edu>
Cc: Denys Vlasenko <vda.linux@googlemail.com>
Cc: Hugh Dickins <hughd@google.com>