usb: gadget: mass_storage: change "ro" writable mode when "cdrom" changes
We do not want to make it possible to change ro flag when cdrom flag is
enabled, yet since cdrom can be set on sysfs, the permissions of the
ro file will not prevent user from modifying ro state.  This commit
changes fsg_store_ro function to check whether cdrom flag is enabled
and fsg_store_cdrom function to change permissions of the ro file.

Signed-off-by: Michal Nazarewicz <mina86@mina86.com>