usb: gadget: mass_storage: fix "ro" permissions when "cdrom" has changed
With change 48a31af74404e6460eabca410bf0b4a625bfd372 it was disallowed
to change "ro" when "cdrom" is turned on. This worked alright when
"cdrom" is set initially, but many downstream kernels allow "cdrom" to
be changed dynamically through sysfs. It was impossible to change "ro"
when "cdrom" is initially turned on, but at runtime turned off.

This commit will fix the issue while also making "ro" read-only when
"cdrom" is turned on dynamically. The read-only file permissions of "ro"
are removed and ensuring "ro" is not written while "cdrom" is turned on
is done in "fsg_store_ro".

Change-Id: I08d08524ef82f620ea6e4a03285f8565fa096ac0